+++ src/featherlight.gallery.js
@@ -50,6 +50,18 @@
 					_super(event);
 
 					self.afterImage(event);
+			},
+			onKeyDown: function(_super, event){
+				var dir = {
+					37: 'previous', /* Left arrow */
+					39: 'next'			/* Rigth arrow */
+				}[event.keyCode];
+				if(dir) {
+					this.$instance.trigger(dir);
+					event.preventDefault();
+				} else {
+					return _super(event);
+				}
 			}
 		};